covid-scrapy/requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
cfscrape==2.1.1
chardet==3.0.4
cloudscraper==1.2.33
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
dateparser==0.7.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
polling==0.3.1
Protego==0.1.16
proxy-middleware==0.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
queuelib==1.5.0
regex==2020.4.4
requests==2.23.0
requests-toolbelt==0.9.1
Scrapy==2.0.1
scrapy-cloudflare-middleware==0.0.1
scrapy-rotating-proxies==0.6.2
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
typing==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.8
w3lib==1.21.0
zope.interface==5.0.2
